# docs/swagger.yaml
info:
  contact: {}
paths:
  /food/addMenu:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      - description: Food ID
        in: body
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增/修改菜單
      tags:
      - menu module
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首頁
  /menu/deleteFood:
    delete:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 食物名稱
        in: body
        name: food_name
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 刪除食物
      tags:
      - menu module
  /order/addOrder:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      - description: 訂單列表
        in: body
        name: order_list
        schema:
          additionalProperties:
            type: integer
          type: object
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增訂單
      tags:
      - order module
  /order/endOrder:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 訂單ID
        in: body
        name: order_id
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 結束訂單
      tags:
      - order module
  /order/getOrder:
    get:
      parameters:
      - description: token
        in: query
        name: token
        type: string
      responses:
        "200":
          description: code", "message", "data []Order"}
          schema:
            type: string
      summary: 取得訂單
      tags:
      - order module
  /rate/addRate:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      - description: 評分
        in: body
        name: rate
        schema:
          type: number
      - description: 評論
        in: body
        name: comment
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 評價添加/修改
      tags:
      - rates module
  /rate/deleteRate:
    delete:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 評價ID
        in: body
        name: rate_id
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 刪除評價
      tags:
      - rates module
  /restaurant/addFavoriteRestaurant:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增最愛餐廳
      tags:
      - restaurant module
  /restaurant/addRestaurant:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      - description: 餐廳名稱
        in: body
        name: name
        schema:
          type: string
      - description: 營業時間
        in: body
        name: business_hours
        schema:
          type: string
      - description: 地址
        in: body
        name: address
        schema:
          type: string
      - description: 電話
        in: body
        name: phone
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增餐廳
      tags:
      - restaurant module
  /restaurant/deleteRestaurant:
    delete:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 刪除餐廳
      tags:
      - restaurant module
  /restaurant/findFavoriteRestaurantList:
    get:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      responses:
        "200":
          description: code", "message", "data []Restaurant"}
          schema:
            type: string
      summary: 查詢最愛餐廳列表
      tags:
      - restaurant module
  /restaurant/findRestaurantByID:
    get:
      parameters:
      - description: 餐廳ID
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code", "message", "data Restaurant"}
          schema:
            type: string
      summary: 透過ID查詢餐廳
      tags:
      - restaurant module
  /restaurant/findRestaurantList:
    get:
      parameters:
      - description: 餐廳名稱
        in: query
        name: name
        type: string
      responses:
        "200":
          description: code", "message", "data []Restaurant"}
          schema:
            type: string
      summary: 查詢餐廳列表
      tags:
      - restaurant module
  /restaurant/findRestaurantRating:
    get:
      parameters:
      - description: 餐廳ID
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code", "message", "data float64", "data rate"}
          schema:
            type: string
      summary: 查詢餐廳評分
      tags:
      - restaurant module
  /restaurant/removeFavoriteRestaurant:
    delete:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 刪除最愛餐廳
      tags:
      - restaurant module
  /restaurant/updateRestaurantInfomation:
    post:
      parameters:
      - description: token
        in: body
        name: token
        schema:
          type: string
      - description: 餐廳ID
        in: body
        name: restaurant_id
        schema:
          type: string
      - description: 餐廳名稱
        in: body
        name: name
        schema:
          type: string
      - description: 營業時間
        in: body
        name: business_hours
        schema:
          type: string
      - description: 地址
        in: body
        name: address
        schema:
          type: string
      - description: 電話
        in: body
        name: phone
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 修改餐廳資訊
      tags:
      - restaurant module
  /user/forget:
    post:
      parameters:
      - description: email信箱
        in: query
        name: email
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 忘記密碼
      tags:
      - user module
  /user/login:
    post:
      parameters:
      - description: email信箱
        in: body
        name: name
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        schema:
          type: string
      responses:
        "200":
          description: code", "message", "token"}
          schema:
            type: string
      summary: 登入帳號
      tags:
      - user module
  /user/register:
    post:
      parameters:
      - description: email信箱
        in: body
        name: name
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        schema:
          type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 註冊新用戶
      tags:
      - user module
  /user/verify:
    post:
      parameters:
      - description: token
        in: query
        name: token
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 驗證帳號
      tags:
      - user module
swagger: "2.0"